Finishing TODO's for installation setup.
I added the Node.js prerequisite and opened to change the SDK installation instructions from Android and Xcode to the three separate OS instructions with links to the Official Documentation. It makes sense to change these since they are essentially just links to NativeScripts documentation and the setup is different for each major OS.  I hope this is alright.