Update tab-view.md (#96)
From personal experience, and confirmed in the nativescript vue slack channel. The event name fired on a tab change is 'tabChange' not selectedIndexChanged.

I'm not sure if the intention is to change the actual event name, but at the moment it is very confusing for anybody reading the docs so I thought it'd be better to follow the current spec.