Add to "## Basic functionality: Add tasks" (#112)
* Add to "## Basic functionality: Add tasks" 

As I was working through the tutorial I could not figure out why my list would not show up. I think it would be wise to make sure it is stated not to forget to set a height on the <ListView> component.

* Fixed a misspelling

I forgot the > add the end of my List View component.